% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst_class.R
\name{analyst_class}
\alias{analyst_class}
\title{Structured ANALYST class}
\description{
This class extends existing classes to provide a unique and structured response from ANALYST API requests.
}
\section{Slots}{

\describe{
\item{\code{content}}{A list containing response content.}

\item{\code{values}}{A data.frame holding the requested data in wide format.}

\item{\code{path}}{A character with URL path of request.}

\item{\code{response}}{An object of class `response`, which inlcudes the entire response from request.}
}}

\keyword{internal}
