% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst_id.R
\name{analyst_id}
\alias{analyst_id}
\title{Create query ID}
\usage{
analyst_id(json = NULL, query_id = FALSE, full_stats = FALSE)
}
\arguments{
\item{json}{Request body in JSON format.}

\item{query_id}{If TRUE, only the query ID is returned.}

\item{full_stats}{Usually only the most basic statistics for a query are pre-computed, 
but if this flag is true, a lot more statistics will be computed beforehand 
(with increased loading time for the query, but potentially decreased loading time for the statistics).}
}
\value{
If \code{query_id = TRUE}, \code{\link{analyst_id}} returns an integer, 
otherwise an object of class \code{analyst_class}, see \code{\link{analyst_class}}.
}
\description{
Create query ID on a given JSON, with the provided filter conditions.
}
\details{
The returned query ID can be used in subsequent API requests to access the data described by this query. 
The queryId is only valid for 6 hours, before it expires and has to be posted again.
}
\seealso{
\code{\link{analyst_response}}
}
