% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst_queries.R
\name{analyst_queries}
\alias{analyst_queries}
\title{Meta data from ANALYST API}
\usage{
analyst_queries(
  id = NULL,
  subquery = c("counterpart", "limit", "benchmarks", "queryId"),
  limit = 10
)
}
\arguments{
\item{id}{ID of the created query to be reused for querying results.}

\item{subquery}{The subquery to be returned.}

\item{limit}{Limit of returned benchmark queries (default: 10), ignored if subquery is not `benchmarks`.}
}
\value{
Returns an object of class \code{analyst_class}.
}
\description{
Function to get additional infornation on query.
}
\seealso{
\code{\link{analyst_response}}
}
