% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst_results.R
\name{analyst_results}
\alias{analyst_results}
\title{Results from ANALYST API}
\usage{
analyst_results(
  id = NULL,
  subquery = c("aggregated", "count", "offers", "percentile", "timeline"),
  variable = NULL,
  aggregation = c("AVG", "MEDIAN", "AVGFG", "STDDEV", "MIN", "MAX", "COUNT"),
  percentile = 50,
  yearparts = 1
)
}
\arguments{
\item{id}{ID of the created query to be reused for querying results.}

\item{subquery}{The subquery to be returned.}

\item{variable}{A target variable for `aggregated`, `percentile` or `timeline` results. If `NULL` and `subquery == 'timeline'`, valid variables for query will be returned.
Ignored in all other subqueries.}

\item{aggregation}{Aggregation mode for subquery `aggregated`, ignored in all other subqueries.}

\item{percentile}{Percentile between 0 and 100 for subquery `percentile`, ignored in all other subqueries.}

\item{yearparts}{Partitioning of timeline with 1 for years, 2 for half-years, 4 for quarters and 12 for months.}
}
\value{
Returns an object of class \code{analyst_class}.
}
\description{
Function to get results from your query ID.
}
\seealso{
, \code{\link{analyst_response}} \code{\link{analyst_id}}
}
