% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_response.R
\name{api_response}
\alias{api_response}
\title{Requests from API}
\usage{
api_response(path = NULL, type = c("GET", "POST", "HEAD"), json = NULL)
}
\arguments{
\item{path}{Path to the desired endpoint.}

\item{type}{One of GET, POST or HEAD.}

\item{json}{Request body in JSON format.}
}
\value{
An object of class \code{\link{api_class}}.
}
\description{
Functions to set up requests from VALUE marketdata API.
}
\details{
These (internal) functions are helpers to start API requests. 
They require valid credentials to access VALUE Marktdaten API. 
See \code{\link{apiur_access}} to set up access.
}
\section{Note}{

\code{\link[httr:content]{httr::content()}} will use `UTF-8` for encoding.
}

\examples{
\dontrun{api_response(path = "status")}
}
\seealso{
\code{\link[httr:GET]{httr::GET()}}, \code{\link[httr:POST]{httr::POST()}}
}
