% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{avm}
\alias{avm}
\title{Request values from AVM indications}
\usage{
avm(
  indication = NULL,
  json = NULL,
  metrics = FALSE,
  details = FALSE,
  market_stats = FALSE,
  comparables = NULL
)
}
\arguments{
\item{indication}{The indication you want to send your request to, e.g. `MARKET_VALUE`. See also \code{\link{avm_indications}} to get all licensed indications.}

\item{json}{Request body in JSON format.}

\item{metrics}{If TRUE, model metrics will be included if available and licensed for chosen indication.}

\item{details}{If TRUE, details will be included if available and licensed for chosen indication.
This is usefull for parallel requests to avoid that function stops on single errors.
Defaults to \code{FALSE}.}
}
\value{
An object of class \code{\link{avm_class}}.
}
\description{
Function to receive values from AVM API.
}
\details{
It returns a list of data frames with all input parameters including input categories 
and all output paramaters including information on JSON output.
Columns with example, minimal or maximal values might also contain dates, that's why they return characters.
}
\examples{
\dontrun{
	
json <- '{
"address": "Heidestraße 8, 10557 Berlin",
"segment": "WHG_K",
"space_living": 100,
"year_of_construction": 1990,
"quality_furnishings": 1
}'
	
avm(indication = 'COMPARATIVE_VALUE', json = json)

}
}
\seealso{
\code{\link{avm_response}}, \code{\link{avm_indications}}, \code{\link{avm_segments}}
}
