% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm_class.R
\name{avm_class}
\alias{avm_class}
\title{Structured AVM class}
\description{
This class extends existing classes to provide a unique and structured response from AVM API requests.
}
\section{Slots}{

\describe{
\item{\code{content}}{A list containing response content.}

\item{\code{values}}{A data.frame holding the requested data in wide format.}

\item{\code{ranges}}{A data.frame holding the requested ranges if available in wide format.}

\item{\code{metrics}}{A data.frame holding metrics data if available in wide format.}

\item{\code{details}}{A data.frame holding details data if available in wide format.}

\item{\code{path}}{A character with URL path of request.}

\item{\code{response}}{An object of class `response`, which inlcudes the entire response from request.}

\item{\code{status}}{An object of class `integer` including \code{status_code} of repsonse.}
}}

\keyword{internal}
