% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm_examples.R
\name{avm_examples}
\alias{avm_examples}
\title{Example values from AVM API}
\usage{
avm_examples()
}
\value{
An object of class \code{data.frame}.
}
\description{
Get example values from example requests.
}
\details{
Based on provided exampleRequests, you will get all output values.
These values contain several formats such as numbers or dates. 
To provide a unique column, we decided to transform all values into characters.
}
\examples{
\dontrun{avm_examples()}
}
\seealso{
\code{\link{avm_response}}
}
