% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm_meta.R
\name{avm_meta}
\alias{avm_meta}
\alias{avm_status}
\alias{avm_endpoints}
\alias{avm_segments}
\title{Meta data from AVM API}
\usage{
avm_status()

avm_endpoints()

avm_segments()
}
\value{
An object of class \code{data.frame}.
}
\description{
Functions to receive meta data from AVM API.
}
\section{Functions}{
\itemize{
\item \code{avm_status()}: Get information on AVM status.

\item \code{avm_endpoints()}: Get information on AVM endpoints with access granted.

\item \code{avm_segments()}: Get segments by AVM endpoint.

}}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{#colnames(avm.endpoints())
}\if{html}{\out{</div>}}
}

\seealso{
\code{\link{avm_response}}
}
