% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm_response.R
\name{avm_response}
\alias{avm_response}
\title{Requests from AVM}
\usage{
avm_response(path = NULL, type = c("GET", "POST", "HEAD"), json = NULL)
}
\arguments{
\item{path}{Path to the desired endpoint.}

\item{type}{One of GET, POST or HEAD.}

\item{json}{Request body in JSON format.}
}
\value{
An object of class \code{\link{avm_class}}.
}
\description{
Functions to set up requests from VALUE market data AVM.
}
\details{
This functions is a helper to start AVM requests. 
It requires valid credentials to access VALUE Marktdaten AVM API. 
See \code{\link{valuer_access}} to set up access.
}
\section{Note}{

\code{\link[httr:content]{httr::content()}} will use `UTF-8` for encoding. 
There is currentlly no option to change this.
}

\examples{
\dontrun{avm_response(path = "status")}
}
\seealso{
\code{\link[httr:GET]{httr::GET()}}, \code{\link[httr:POST]{httr::POST()}}
}
