% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm_specifications.R
\name{avm_specifications}
\alias{avm_specifications}
\title{Specifications from AVM API}
\format{
An object of class \code{list()} including \code{data.frames}
}
\usage{
avm_specifications(endpoint = NULL, segments = NULL, language = c("EN", "DE"))
}
\arguments{
\item{endpoint}{A vector of indication endpoints with \code{specification = T}. See also \code{\link{avm_endpoints}} to get all valid endpoints.}

\item{segments}{A vector of segments. See also \code{\link{avm_segments}} to get all valid segments.}

\item{language}{Specification of english or german language used to format character columns with englisch settings as DEFAULT. If set to `DE`, big (`.`) and decimal (`,`) marks will be set and dates to `dd.mm.yyy`.}
}
\description{
Function to receive information on valid specifications of AVM API.
}
\details{
It returns a list of data frames with all input parameters including input categories 
and all output paramaters including information on JSON output.
Columns with example, minimal or maximal values might also contain dates, that's why they return characters.
}
\seealso{
\code{\link{avm_response}}, \code{\link{avm_endpoints}}, \code{\link{avm_segments}}
}
