% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_account}
\alias{calc_account}
\title{Calculates the account}
\usage{
calc_account(spot, fee, barrier)
}
\arguments{
\item{spot}{\code{numeric} vector with the VA reference fund values}

\item{fee}{\code{numeric} scalar with the fee}

\item{barrier}{\code{numeric} scalar with the state-dependent barrier}
}
\description{
Calculates the account
}

