% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exotic_pricer.R
\docType{class}
\name{cash_flow}
\alias{cash_flow}
\title{Class providing a cash flow object with methods to get/set the cash flow amount.}
\format{\code{\link{R6Class}} object.}
\usage{
cash_flow
}
\value{
Object of \code{\link{R6Class}} with methods to get/set the cash flow amount.
}
\description{
Class providing a cash flow object with methods to get/set the cash flow amount.
}
\section{Fields}{

\describe{
\item{\code{amount}}{Stores the amount of the cash flow.}

\item{\code{time_index}}{Stores the time index of the cash flow.}
}}
\section{Methods}{

\describe{
  \item{\code{get_amount}}{get method for the amount of the cash flow}

  \item{\code{set_amount}}{set method for the amount of the cash flow }
  }
}
\examples{
cash <- cash_flow$new(10, 50.89)
cash$get_amount()
cash$set_amount(60.00)
cash$get_amount()
}
\keyword{datasets}

