% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exotic_pricer.R
\docType{class}
\name{cash_flows}
\alias{cash_flows}
\title{Class providing a cash flow object with methods to get/set the cash flow amounts.}
\format{\code{\link{R6Class}} object.}
\usage{
cash_flows
}
\value{
Object of \code{\link{R6Class}} with methods to get/set the cash flow amounts.
}
\description{
Class providing a cash flow object with methods to get/set the cash flow amounts.
}
\section{Fields}{

\describe{
\item{\code{amounts}}{\code{numeric} Stores the amounts of the cash flows.}

\item{\code{time_index}}{\code{\link{timeDate}} Stores the time-line of the cash flows.}
}}
\section{Methods}{

\describe{
  \item{\code{get_amounts}}{get method for the amounts of the cash flow. Returns a \code{numeric} scalar or vector}

  \item{\code{set_amounts}}{set method for the cash flow amounts }
  }
}
\examples{
tt <- timeDate::Sys.timeDate()
cash <- cash_flows$new(tt, 50.89)
cash$get_amounts()
cash$set_amounts(60.00)
cash$get_amounts()
}
\keyword{datasets}

