% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptionPricing.R
\docType{class}
\name{constant_parameters}
\alias{constant_parameters}
\title{Class providing a constant parameter object.}
\format{\code{\link{R6Class}} object.}
\usage{
r <- constant_parameters$new(0.01/365)
r$integral(timeDate("2016-07-09"), timeDate("2017-07-09"))
r$integral_square(timeDate("2016-07-09"), timeDate("2017-07-09"))
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class providing a constant parameter object.
}
\section{Fields}{

\describe{
\item{\code{constant}}{(\code{private}) A \code{numeric} which stores the constant parameter.}

\item{\code{constant_square}}{(\code{private} A \code{numeric} which stores the square of the constant parameter}
}}
\section{Methods}{

\describe{
  \item{\code{integral} (\code{public})}{Calculates the integral given the initial and final times. The arguments are two \code{\link{timeDate}} object with the initial  and final times. It returns a \code{numeric} scalar with the integral}

  \item{\code{integral_square} (\code{public})}{Calculates the integral of the squared constant parameter given the initial and final times. The arguments are two \code{\link{timeDate}} object with the initial  and final times. It returns a \code{numeric} scalar with the integral}
  }
}
\keyword{datasets}

