% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_gatherer.R
\docType{class}
\name{data_gatherer}
\alias{data_gatherer}
\title{Simple data gatherer}
\format{\code{\link{R6Class}} object.}
\usage{
data_gatherer
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class which defines a simple data gatherer to hold estimates
calculated in a loop.
}
\section{Methods}{

\describe{
  \item{\code{new}}{Constructor method}
  \item{\code{dump_result}}{Saves the argument \code{result} which is
  a numeric scalar}
  \item{\code{get_results}}{Returns a numeric vector with the point
  estimates.}
}
}

\keyword{datasets}
