% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_for_data_objects.R
\docType{data}
\name{financials_BZ2016bis}
\alias{financials_BZ2016bis}
\title{BZ2016bis financial processes}
\format{A list with elements:
\describe{
\item{[[1]]}{List of parameters for \code{\link[yuima]{simulate}}}
\item{[[2]]}{List of parameters for \code{\link{setModel}}}
\item{[[3]]}{Vector with indices indicating the log price
in solve.variable \code{\link{setModel}}}
}}
\usage{
financials_BZ2016bis
}
\description{
List of parameters to initialize a \code{\link{va_sde_engine3}} object to
simulate the  log price and volatility processes which
follow the stochastic differential equations specified
in BMOP2011 -  See \bold{References}. The interest rate is constant
with default value 0.03.
}
\examples{
#Sets the interest rate to 2\%
financials_BZ2016bis[[1]]$r <- 0.02

}
\references{
\enumerate{
\item{[BMOP2011]}{
 \cite{Bacinello A.R., Millossovich P., Olivieri A. e Pitacco E.
       "Variable annuities: a unifying valuation approach."
        In: Insurance: Mathematics and Economics 49 (2011), pp. 285-297.}
      }
 }
}
\keyword{datasets}
