% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_products.R
\docType{class}
\name{GMAB}
\alias{GMAB}
\title{Variable Annuity with GMAB guarantee}
\format{\code{\link{R6Class}} object.}
\usage{
#Sets up the payoff as a roll-up of premiums with roll-up rate 1%

rate <- constant_parameters$new(0.01)

premium <- 100
rollup <- payoff_rollup$new(premium, rate)

#Five years time-line
times <- timeDate::timeSequence(from="2016-01-01", to="2020-12-31")

# A constant fee of 2% per year (365 days)
fee <- constant_parameters$new(0.02, 365)

#Barrier for a state-dependent fee. The fee will be applied only if
#the value of the account is below the barrier
barrier <- 200

#Withdrawal penalty applied in case the insured surrenders the contract.
penalty <- 0.01

VA_GMAB <- GMAB$new(rollup, times, fee, barrier, penalty)
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class for a  VA product object with Guaranteed Minimum Accumulation Benefit (GMAB). \cr It inherits from \code{\link{va_product}}.
It supports a simple state-dependent fee structure with a single barrier.
See \bold{References} for a description of variable annuities life insurance products, their guarantees and fee structures.
}
\section{Fields}{

\describe{
\item{\code{the_payoff}}{(\code{private}) A \code{\link{payoff_guarantee}} object which stores the type of payoff
(e.g: roll-up, ratchet, etc).}

\item{\code{the_fee}}{(\code{private}) A positive scalar with the annual VA contract fee.}

\item{\code{the_barrier}}{(\code{private}) A positive scalar with the barrier for state-dependent fees. The fee will be applied only if the value of the account is below the barrier.}

\item{\code{the_penalty}}{(\code{private}) A scalar with the withdrawal penalty in case the insured surrenders the contract.
Must be between 0 and 1.}

\item{\code{times_yrs}}{(\code{private}) A numeric vector with the product time-line in fraction of years}
}}
\section{Methods}{

 \describe{
   \item{\code{new}}{Constructor method}
   \item{\code{get_times}}{get method for the product time-line. Returns a \code{\link{timeDate}} object}
   \item{\code{times_in_yrs}}{get method for the product time-line in fraction of year}
   \item{\code{max_number_cfs}}{ returns an \code{integer} with the maximun number of cash flows the product can generate}
   \item{\code{cash_flow_times}}{retuns a \code{\link{timeDate}} object with the possible cash flow times. Takes as argument a \code{\link{timeDate}} object with the time of death}
   \item{\code{survival_benefit_times}}{retuns a \code{numeric} vector with the survival benefit time indexes.}
   \item{\code{surrender_times}}{retuns a \code{numeric} vector with the surrender time indexes. Takes as argument a string with the frequency of the decision if surrendering the contract,  e.g. "3m"  corresponds to a surrender decision taken every 3 months.}
   \item{\code{death_benefit_times}}{retuns a \code{numeric} vector with the death benefit time indexes if a death benefit is present in the rider.}
   \item{\code{cash_flows}}{returns a \code{numeric} vector with the cash flows of the product. It takes as argument \code{spot_values} a \code{numeric} vector which holds the values of the underlying fund, \code{death_time} the  index of the \code{times} vector corresponding to the time of death and the last index in case the insured survives.}
   \item{\code{get_premium}}{Returns the premium as non negative scalar}
}
}
\references{
\enumerate{
\item{\cite{ Bacinello A.R., Millossovich P., Olivieri A., Pitacco  E.,
"Variable annuities: a unifying valuation approach." In: Insurance: Mathematics and Economics 49 (2011), pp. 285-297.
}}
\item{\cite{Bernard C., Hardy M. and Mackay A. "State-dependent fees for variable
 annuity guarantees." In: Astin Bulletin 44 (2014), pp. 559-585.}}
 }
}
\keyword{datasets}

