% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeham}
\alias{makeham}
\title{Makeham's intensity of mortality}
\usage{
makeham(t, x, A, B, c)
}
\arguments{
\item{t}{time as numeric scalar}

\item{x}{age as numeric scalar}

\item{A}{numeric scalar}

\item{B}{numeric scalar}

\item{c}{numeric scalar}
}
\description{
Makeham's intensity of mortality
}
