% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_gatherer.R
\docType{class}
\name{mc_gatherer}
\alias{mc_gatherer}
\title{Monte Carlo gatherer}
\format{\code{\link{R6Class}} object.}
\usage{
the_gatherer <- mc_gatherer$new()
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class which defines a gatherer for the Monte Carlo simulated
values. It has methods to return the Monte Carlo estimate and Monte Carlo
Standard Error of the estimate as well as a convergence table.
}
\section{Methods}{

\describe{
  \item{\code{new}}{Constructor method}
  \item{\code{dump_results}}{Saves the argument \code{result} which is
  a numeric vector of simulated values}
  \item{\code{get_results}}{Returns the  Monte Carlo estimate and the
  (estimated) Monte Carlo Standard Error of the estimate}
  \item{\code{convergence_table}}{Returns the convergence table}
}
}
\keyword{datasets}

