% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_for_data_objects.R
\docType{data}
\name{mortality_BBM2010}
\alias{mortality_BBM2010}
\title{BBM2010 demographic processes}
\format{A list with elements:
\describe{
\item{[[1]]}{List of parameters for \code{\link[yuima]{simulate}}}
\item{[[2]]}{List of parameters for \code{\link{setModel}}}
\item{[[3]]}{Vector with indices indicating the intensity of mortality
in solve.variable \code{\link{setModel}}}
}}
\usage{
mortality_BBM2010
}
\description{
List of parameters to initialize a va_sde_engine object to
simulate the intensity of mortality process
according to the stochastic differential equation specified
in BBM2010 -  See \bold{References}.
}
\references{
\enumerate{
\item{[BBM2010]}{
 \cite{Bacinello A.R., Biffis E. e Millossovich P.
       "Regression-based algorithms
        for life insurance contracts with surrender guarantees".
        In: Quantitative Finance 10.9 (2010), pp. 1077-1090.}
       }
 }
}
\keyword{datasets}
