% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_for_data_objects.R
\docType{data}
\name{mortality_BMOP2011}
\alias{mortality_BMOP2011}
\title{BMOP2011 demographic processes}
\format{A list with elements:
\describe{
\item{[[1]]}{List of parameters for \code{\link[yuima]{simulate}}}
\item{[[2]]}{List of parameters for \code{\link{setModel}}}
\item{[[3]]}{Vector with indices indicating the intensity of mortality
in solve.variable \code{\link{setModel}}}
}}
\usage{
mortality_BMOP2011
}
\description{
List of parameters to initialize a va_sde_engine object to
simulate the intensity of mortality process
according to the stochastic differential equation specified
in BMOP2011 - See \bold{References}.
}
\references{
\enumerate{
\item{[BMOP2011]}{
 \cite{Bacinello A.R., Millossovich P., Olivieri A. e Pitacco E.
       "Variable annuities: a unifying valuation approach."
        In: Insurance: Mathematics and Economics 49 (2011), pp. 285-297.}
      }
 }
}
\keyword{datasets}
