% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mu}
\alias{mu}
\title{Weibull intensity of mortality}
\usage{
mu(t, x, c1, c2)
}
\arguments{
\item{t}{time as numeric scalar}

\item{x}{age as numeric scalar}

\item{c1}{numeric scalar}

\item{c2}{numeric scalar}
}
\description{
Weibull intensity of mortality
}
