% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\docType{class}
\name{payoff_call}
\alias{payoff_call}
\title{Call option payoff}
\format{\code{\link{R6Class}} object.}
\usage{
payoff_call
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class providing a call option payoff.
}
\section{Methods}{

\describe{
 \item{\code{new}}{Constructor method}
 \item{\code{get}}{Get the strike price}
 \item{\code{set}}{Set the stike price}
 \item{\code{payoff}}{Calculates the payoff given the argument
  \code{spot} which is a \code{numeric} vector or scalar with
  the spot prices}
}
}
\examples{
\dontrun{
callpayoff <- payoff_call$new(100)

callpayoff$payoff(c(130, 120))

#The [] operator was overridden to calculate the payoff
#with a function style:

callpayoff[c(130, 120)]
}
}
\keyword{datasets}

