% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sq}
\alias{sq}
\title{Square root utility function}
\usage{
sq(x)
}
\arguments{
\item{x}{numeric scalar}
}
\description{
Takes square root if positive otherwise returns zero.
To be used with mean reverting squared root processes (CIR SDE)
}
