% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_pricer.R
\docType{class}
\name{va_bs_engine}
\alias{va_bs_engine}
\title{Variable Annuity  pricing engine with GBM}
\format{\code{\link{R6Class}} object.}
\usage{
contract <- GMAB$new(rollup, times, age, fee, barrier, penalty)

r <- constant_parameters$new(0.01)
spot <- 100
vol <- constant_parameters$new(0.2)
div <- constant_parameters$new(0.0)

gatherer_mean <- statistics_mean$new()
no_of_paths <- 1e4

engine <- va_bs_engine$new(contract, r, c1=90.43, c2=10.36, spot, volatility=vol, dividends=div)

#Estimates the contract value by means of the static approach.

engine$do_static(gatherer_mean, no_of_paths)
gatherer_mean$get_results()

#Estimates the contract value by means of the mixed approach.
#To compare with the static approach we won't simulate the underlying
#fund paths again.

gatherer_mean_2 <- statistics_mean$new()

engine$do_mixed(gatherer_mean_2, no_of_paths, degree = 3, freq = "3m", simulate = FALSE)
gatherer_mean_2$get_results()
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class providing a variable annuity pricing engine with the underlying
reference risk neutral fund modeled as a Geometric Brownian Motion and the
intensity of mortality  modeled by the Weibull intensity of mortality.
It implements the static approach to estimate the value of a VA
contract by means of the Monte Carlo method and the mixed approach by means
of Least Squares Monte Carlo.\cr
See \bold{References} for a description of the mixed and static approaches and
Least Squares Monte Carlo.
}
\section{Fields}{

\describe{
\item{\code{the_product}}{(\code{private}) A \code{\link{va_product}} object with
the VA contract.}

\item{\code{times}}{(\code{private}) \code{\link{timeDate}} object with the product time-line}

\item{\code{r}}{(\code{private}) \code{\link{parameters}} object with the risk-neutral interest rate}

\item{\code{drifts}}{(\code{private}) \code{numeric} vector with the drifts of the underlying
fund.}

\item{\code{standard_deviations}}{(\code{numeric}) vector with the standard_deviations of the
underlying fund.}

\item{\code{spot}}{(\code{numeric}) numeric scalar with the initial value of the underlying fund}

\item{\code{mu_integrals}}{(\code{numeric}) vector with the integrals along the product
time-line of the intensity of mortality}

\item{\code{fund}}{(\code{matrix}) with  simulated paths of the underlying fund}

\item{\code{discounts}}{(\code{numeric}) vector with the discount factors}

\item{\code{mu_1,mu_2}}{(\code{numeric}) scalars with the parameters for the Weibull
intensity of mortality}
}}
\section{Methods}{

\describe{
 \item{\code{new}}{Constructor method}
 \item{\code{get_mortality}}{Returns the integrals of the intensity of
 mortality along the product time-line.}
 \item{\code{death_time}}{Returns the time of death index. If the
 death doesn't occur during the product time-line it returns the
 last index of the product time-line.}
 \item{\code{simulate_financial_paths}}{Simulates \code{npaths} paths
 of the underlying fund of the VA contract and the discount factors (interest
  rate). It saves the paths into the \code{fund} matrix and
  the \code{discounts} vector.}
 \item{\code{get_fund}}{Gets the \code{i}-th path of the underlying fund where
  \code{i} goes from 1 to \code{npaths}.}
 \item{\code{do_static}}{Estimates the VA contract value by means of
 the static approach (Monte Carlo), see \bold{References}. It takes as
 arguments:
  \describe{
    \item{\code{the_gatherer}}{\code{\link{gatherer}} object to hold
    the point estimates}
    \item{\code{npaths}}{positive integer with the number of paths to
    simulate}
    \item{\code{simulate}}{boolean to specify if the paths should be
    simulated from scratch, default is TRUE.}
  }
 }
 \item{\code{do_mixed}}{Estimates the VA contract by means of
 the mixed approach (Least Squares Monte Carlo), see \bold{References}.
 It takes as arguments:
  \describe{
   \item{\code{the_gatherer}}{\code{\link{gatherer}} object to hold
    the point estimates}
    \item{\code{npaths}}{positive integer with the number of paths to
    simulate}
    \item{\code{degree}}{positive integer with the maximum degree of
    the weighted Laguerre polynomials used in the least squares by LSMC}
    \item{\code{freq}}{string which contains the frequency of the surrender
    decision. The default is \code{"3m"} which corresponds to deciding every
    three months if surrendering the contract or not.}
    \item{\code{simulate}}{boolean to specify if the paths should be
    simulated from scratch, default is TRUE.}
  }
 }
 \item{\code{get_discount}}{Arguments are \code{i,j}.
 Gets the \code{j}-th discount factor corresponding to the \code{i}-th
 simulated path of the discount factors.}
 \item{\code{bases}}{Interface to get Laguerre polynomials of  state variables.
 It is implemented in sub-classes since state variables may differ depending
 on the dynamics of financial and mortality processes.
 For example if we're modeling just the undelying fund as a
 stochastic process (all rest deterministic), the state variable
 process will be just the fund, no need to include interest rate or mortality.
 \cr
 Arguments are:
 \describe{
  \item{\code{paths}}{numeric vector of indexes of the financial path processes}
  \item{\code{time}}{numeric scalar with the time index}
  \item{\code{degree}}{positive scalar with the max degree of the Laguerre
 polynomials}
  }
 }
}
}
\references{
\enumerate{
 \item{\cite{ Bacinello A.R., Millossovich P., Olivieri A., Pitacco  E.,
 "Variable annuities: a unifying valuation approach."
 In: Insurance: Mathematics and Economics 49 (2011), pp. 285-297.
}}
 \item{\cite{Longstaff F.A. e Schwartz E.S. Valuing american options by simulation:
 a simple least-squares approach. In: Review of Financial studies
 14 (2001), pp. 113-147}}
 }
}
\keyword{datasets}

