% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{class}
\name{vanilla_option}
\alias{vanilla_option}
\title{Vanilla option}
\format{\code{\link{R6Class}} object.}
\usage{
vanilla_option
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class for a vanilla option supporting different types of payoffs
}
\section{Methods}{

\describe{
 \item{\code{new}}{Constructor method which takes as argument a
  \code{the_payoff} a \code{payoff} object with the payoff and
  \code{expiry} a positive scalar with the expiry time.}
 \item{\code{get_expiry}}{get method for the expiry time.}
 \item{\code{option_payoff}}{returns a \code{numeric} vector with the
  payoff. It takes as argument \code{spot}
  a \code{numeric} vector with the values of the underlying asset
  this method will calculate the payoff from.}
}
}
\examples{
\dontrun{
#Defines the payoff of a call option
call_payoff <- payoff_call$new(100)
#Defines the call option
call_option <- vanilla_option$new(the_payoff = call_payoff, expiry = 1)
#Calculates the payoff of the call corresponding to two spot prices
call_option$option_payoff(c(120, 130))
}
}
\keyword{datasets}

