% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yr_fractions}
\alias{yr_fractions}
\title{Normalizes a timeDate sequence into year fractions}
\usage{
yr_fractions(times)
}
\arguments{
\item{times}{A \code{\link{timeDate}} sequence}
}
\description{
Normalizes a timeDate sequence into year fractions
}
