% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_account}
\alias{calc_account}
\title{Calculates the account}
\usage{
calc_account(spot, ben, fee, barrier, penalty)
}
\arguments{
\item{spot}{\code{numeric} vector with the VA reference fund values}

\item{ben}{\code{numeric} vector with the living benefit cash flow}

\item{fee}{\code{numeric} scalar with the fee}

\item{barrier}{\code{numeric} scalar with the state-dependent barrier}

\item{penalty}{\code{numeric} vector with the surrender penalty}
}
\description{
Calculates the account
}
