% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exotic_pricing_engine.R
\docType{class}
\name{exotic_engine}
\alias{exotic_engine}
\title{Pricing engine of a path dependent derivative}
\format{\code{\link{R6Class}} object.}
\usage{
exotic_engine
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Base class which defines the interface of a pricing engine of
a path dependent european derivative product. This class should
be inherited by subclasses implementing its methods and not be
instantiated directly.
}
\section{Methods}{

\describe{
 \item{\code{new}}{Constructor method which takes as argument a
  \code{\link{path_dependent}} object with the product and a
  \code{parameters} object with the interest rate}
 \item{\code{get_one_path}}{Returns a \code{numeric}
  vector with a simulated path of the product underlying asset. This
  method needs to be implemented by subclasses.}
 \item{\code{run_simulation}}{Runs the Monte Carlo
  simulation and stores the results in a \code{gatherer} object.
  Takes as arguments a \code{gatherer} object to store
  the results and an \code{integer} scalar with the number of paths
  to simulate.}
 \item{\code{discount_one_path}}{Discounts a simulated path of the
  underlying asset. It takes as argument \code{spot_values} a
  \code{numeric} vector of simulated values of the underlying asset,
  calculates the cash flows and takes their present value given the
  discount factors. Returns a \code{numeric} scalar with the present
  value of the discounted product cash flows.}
 \item{\code{discounts}}{\code{\link{R6Class}} active binding which
  calculates the discount factors.}
}
}
\keyword{datasets}

