% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exotic_pricer.R
\docType{class}
\name{exotic_engine}
\alias{exotic_engine}
\title{Class which defines  the interface of a pricing engine of a path dependent derivative}
\format{\code{\link{R6Class}} object.}
\usage{
exotic_engine
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Base class which defines the interface of a pricing engine of a path dependent derivative
product. This class should be inherited by subclasses implementing its methods.
Objects of this class should not be instantiated.
}
\section{Fields}{

\describe{
\item{\code{the_product}}{A \code{\link{path_dependent}} object with the product the engine will price.}

\item{\code{r}}{A \code{\link{parameters}} object with the interest rate.}
}}
\section{Methods}{

\describe{
  \item{\code{new}}{Constructor method which takes as argument a \code{\link{path_dependent}} object with the product and a \code{\link{parameters}} object with the interest rate}
  \item{\code{get_one_path}}{Returns a \code{numeric} vector with a simulated path of the product underlying asset}
  \item{\code{run_simulation}}{Runs the Monte Carlo simulation and stores the results in a \code{\link{gatherer}} object. Takes as arguments a \code{\link{gatherer}} object to store the results and an \code{integer} scalar with the number of paths to simulate.}
  \item{\code{discount_one_path}}{discounts a cash flow path}
  \item{\code{discounts}}{\code{\link{R6Class}} active binding which calculates the discount factors.}
  }
}
\keyword{datasets}

