% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\docType{class}
\name{payoff_guarantee}
\alias{payoff_guarantee}
\title{Generic guarantee payoff class}
\format{\code{\link{R6Class}} object.}
\usage{
payoff_guarantee
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class providing an interface for guarantee payoff objects.
This class shouldn't be instantiated but used as base class
for more specialized implementations such as a roll-up or
ratchet payoff classes.
}
\section{Methods}{

 \describe{
  \item{\code{new} (\code{public})}{Initialize method.
  The argument is a non negative scalar with the premium.}
  \item{\code{set_premium} (\code{public})}{Stores the
  premium in a private field. The argument is a non negative scalar}
  \item{\code{get_premium} (\code{public})}{Returns the premium
  as non negative scalar}
  \item{\code{get_payoff} (\code{public})}{Gets a zero payoff in this
   base class.The arguments are a \code{numeric} vector with
   the amounts and a vector of \code{\link{timeDate}}
   objects to calculate the payoff}
}
}

\keyword{datasets}
