% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_pricer.R
\docType{class}
\name{va_engine}
\alias{va_engine}
\title{Generic Variable Annuity  pricer engine}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
va_engine
}
\description{
Class providing an interface for a generic VA pricer engine.
See \bold{References} for a description of variable annuities life insurance products, their guarantees and fee structures.
}
\keyword{datasets}

