% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuer_access.R
\name{valuer_access}
\alias{valuer_access}
\title{Set up connection to API}
\usage{
valuer_access(
  avm_url = Sys.getenv("VALUER_AVM_URL"),
  avm_username = Sys.getenv("VALUER_AVM_USER"),
  avm_password = Sys.getenv("VALUER_AVM_PW"),
  api_url = Sys.getenv("VALUER_ANALYST_URL"),
  api_username = Sys.getenv("VALUER_ANALYST_USER"),
  api_password = Sys.getenv("VALUER_ANALYST_PW")
)
}
\arguments{
\item{avm_url}{A valid url from which you want to access our AVM with Sys.getenv("VALUER_AVM_URL") as default.}

\item{avm_username}{A character vector for username with Sys.getenv("VALUER_AVM_USER") as default.}

\item{avm_password}{A character vector for password with Sys.getenv("VALUER_AVM_PW") as default.}

\item{analyst_url}{A valid url from which you want to access our AVM with Sys.getenv("VALUER_ANALYST_URL") as default.}

\item{analyst_username}{A character vector for username with Sys.getenv("VALUER_ANALYST_USER") as default.}

\item{analyst_password}{A character vector for password with Sys.getenv("VALUER_ANALYST_API_PW") as default.}
}
\description{
Using \code{valuer_access()} you can connect to any valid and licensed API from VALUE Marktdaten.
}
\section{Note}{
 You must have a valid license.
}

\examples{
\dontrun{valuer_access()}
}
\seealso{
\code{\link{avm_response}}, \code{\link[httr:GET]{httr::GET()}}
}
